feat(nav): refactor nav item list to enable shadowing Refactor building of the nav item list to standalone file for override Enables alternate mechanisms for building the nav item list
