+++ packages/gatsby-theme-carbon/src/components/LeftNav/LeftNavItemProvider.js
@@ -1,0 +1,24 @@
+import { useStaticQuery, graphql } from 'gatsby';
+
+export function useNavItems() {
+  const {
+    allNavItemsYaml: { edges },
+  } = useStaticQuery(graphql`
+    query LEFT_NAV_QUERY {
+      allNavItemsYaml {
+        edges {
+          node {
+            title
+            pages {
+              title
+              path
+            }
+          }
+        }
+      }
+    }
+  `);
+
+  const navItems = edges.map(({ node }) => node);
+  return navItems;
+}
